+++ src/prelude.rs
@@ -20,7 +20,11 @@ pub use sea_orm::{
 pub use serde_json::json as data;
 
 #[cfg(all(feature = "auth_jwt", feature = "with-db"))]
-pub use crate::controller::middleware::auth;
+pub use crate::controller::extractor::auth;
+pub use crate::controller::extractor::{
+    shared_store::SharedStore,
+    validate::{JsonValidate, JsonValidateWithMessage},
+};
 #[cfg(feature = "with-db")]
 pub use crate::model::{query, Authenticable, ModelError, ModelResult};
 pub use crate::{